requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
pydantic
pandas
requests
python-dotenv
pytest
httpx